Retrieve updates from dev

+++ libs/my_c/src/array/my_str_to_word_array.c
@@ -12,7 +12,7 @@
 
 #include "../../include/my.h"
 
-static bool is_separator(char *str, int64_t index,
+static bool is_separator(char *str, size_t index,
     char *separator, separator_type_t type)
 {
     switch (type) {
@@ -33,13 +33,13 @@ static bool is_separator(char *str, int64_t index,
     return false;
 }
 
-int64_t count_words(char *str, char *separator, separator_type_t type)
+size_t count_words(char *str, char *separator, separator_type_t type)
 {
-    int64_t nb_words = 0;
+    size_t nb_words = 0;
     bool add_word = true;
-    int64_t len_separator = my_strlen(separator) - 1;
+    size_t len_separator = my_strlen(separator) - 1;
 
-    for (int64_t index = 0; str[index] != '\0'; index++) {
+    for (size_t index = 0; str[index] != '\0'; index++) {
         if (!is_separator(str, index, separator, type) && add_word) {
             nb_words++;
             add_word = false;
@@ -53,10 +53,10 @@ int64_t count_words(char *str, char *separator, separator_type_t type)
     return nb_words;
 }
 
-static void move_save(char *str, int64_t *saved_index,
+static void move_save(char *str, size_t *saved_index,
     char *separator, separator_type_t type)
 {
-    int64_t len_separator = my_strlen(separator) - 1;
+    size_t len_separator = my_strlen(separator) - 1;
 
     for (; str[*saved_index] != '\0' &&
     is_separator(str, *saved_index, separator, type); (*saved_index)++)
@@ -64,12 +64,12 @@ static void move_save(char *str, int64_t *saved_index,
             *saved_index + len_separator : *saved_index);
 }
 
-int64_t count_letters(char *str, int64_t *saved_index,
+size_t count_letters(char *str, size_t *saved_index,
     char *separator, separator_type_t type)
 {
-    int64_t nb_letters = 0;
+    size_t nb_letters = 0;
 
-    for (int64_t index = *saved_index; str[index] != '\0'
+    for (size_t index = *saved_index; str[index] != '\0'
     && !is_separator(str, index, separator, type); index++)
         nb_letters++;
     *saved_index = *saved_index + nb_letters;
@@ -87,12 +87,12 @@ int64_t count_letters(char *str, int64_t *saved_index,
  */
 char **my_str_to_word_array(char *str, char *separator, separator_type_t type)
 {
-    int64_t nb_words = count_words(str, separator, type);
+    size_t nb_words = count_words(str, separator, type);
     char **array = my_malloc(sizeof(char *) * (nb_words + 1));
-    int64_t saved_index = 0;
-    int64_t nb_letters = 0;
+    size_t saved_index = 0;
+    size_t nb_letters = 0;
 
-    for (int64_t index = 0; index < nb_words; index++) {
+    for (size_t index = 0; index < nb_words; index++) {
         move_save(str, &saved_index, separator, type);
         nb_letters = count_letters(str, &saved_index, separator, type);
         array[index] = my_strndup(str + saved_index - nb_letters, nb_letters);
