+++ libs/my_c/src/params/my_rev_params.c
@@ -20,7 +20,7 @@
  */
 void my_rev_params(int argc, char **argv)
 {
-    for (int i = argc; i > 0; i--) {
+    for (ssize_t i = argc; i > 0; i--) {
         my_putstr(argv[i - 1]);
         my_putstr("\n");
     }
