+++ libs/my_c/src/string/create_str/flag_lowercase_a.c
@@ -103,7 +103,7 @@ static void precise_zero(char *float_nb, int precision)
         my_strcat(float_nb, "0x0p+0");
     } else {
         my_strcat(float_nb, "0x0.");
-        for (int i = 0; i < precision; i++)
+        for (size_t i = 0; i < precision; i++)
             my_strcat(float_nb, "0");
         my_strcat(float_nb, "p+0");
     }
