+++ libs/my_c/src/string/my_strupcase.c
@@ -20,7 +20,7 @@
  */
 char *my_strupcase(char *str)
 {
-    for (int i = 0; str[i] != '\0'; i++) {
+    for (size_t i = 0; str[i] != '\0'; i++) {
         if (str[i] >= 97 && str[i] <= 122)
             str[i] = str[i] - 32;
     }
